move commands to `mpg` subcomand
